+++ AddOns/LiteButtonAuras/CHANGELOG.md
@@ -1,6 +1,9 @@
 # LiteButtonAuras
 
-## [10.2.0-6](https://github.com/xod-wow/LiteButtonAuras/tree/10.2.0-6) (2024-01-09)
-[Full Changelog](https://github.com/xod-wow/LiteButtonAuras/compare/10.2.0-5...10.2.0-6) [Previous Releases](https://github.com/xod-wow/LiteButtonAuras/releases)
+## [10.2.0-7](https://github.com/xod-wow/LiteButtonAuras/tree/10.2.0-7) (2024-01-16)
+[Full Changelog](https://github.com/xod-wow/LiteButtonAuras/compare/10.2.0-6...10.2.0-7) [Previous Releases](https://github.com/xod-wow/LiteButtonAuras/releases)
 
-- Fix the aura display width so it doesn't sometimes bug out  
+- Add slash commands for anchors  
+- More options GUI work  
+- Show taunts from other players/pets on player taunt ability  
+- Only match long friendly buffs from others